implementations moved to struct declaration files
