--- defuse/src/contract/accounts/account/entry/implementations/mod.rs
@@ -1,4 +1,0 @@
-/// There are  copies of legacy implementations of accounts, so we can handle
-/// all mutations same way as they happened before
-mod v0;
-mod v1;
