+++ defuse/src/contract/accounts/account/entry/mod.rs
@@ -1,4 +1,3 @@
-mod implementations;
 mod v0;
 mod v1;
