+++ defuse/src/contract/accounts/account/entry/v0.rs
@@ -1,11 +1,18 @@
 use defuse_bitmap::{U248, U256};
-use defuse_core::{Nonces, crypto::PublicKey};
 use defuse_near_utils::NestPrefix;
 use impl_tools::autoimpl;
 use near_sdk::{
-    near,
+    AccountIdRef, near,
     store::{IterableSet, LookupMap},
 };
+
+use defuse_core::{
+    Nonces, Result,
+    accounts::{AccountEvent, PublicKeyEvent},
+    crypto::PublicKey,
+    events::DefuseEvent,
+};
+use std::borrow::Cow;
 
 use crate::contract::accounts::{
     Account, AccountState, MaybeLegacyNonces,
@@ -50,5 +57,93 @@ impl From<AccountV0> for Account {
             state,
             prefix,
         }
+    }
+}
+
+/// Legacy implementation of [`AccountV0`]
+impl AccountV0 {
+    #[cfg(test)]
+    #[inline]
+    pub fn new<S>(prefix: S, me: &AccountIdRef) -> Self
+    where
+        S: near_sdk::IntoStorageKey,
+    {
+        let prefix = prefix.into_storage_key();
+
+        Self {
+            nonces: Nonces::new(LookupMap::new(
+                #[allow(deprecated)]
+                prefix.as_slice().nest(AccountPrefix::LegacyNonces),
+            )),
+            implicit_public_key_removed: !me.get_account_type().is_implicit(),
+            public_keys: IterableSet::new(prefix.as_slice().nest(AccountPrefix::PublicKeys)),
+            state: AccountState::new(prefix.as_slice().nest(AccountPrefix::State)),
+            prefix,
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn add_public_key(&mut self, me: &AccountIdRef, public_key: PublicKey) -> bool {
+        if !self.maybe_add_public_key(me, public_key) {
+            return false;
+        }
+
+        DefuseEvent::PublicKeyAdded(AccountEvent::new(
+            Cow::Borrowed(me),
+            PublicKeyEvent {
+                public_key: Cow::Borrowed(&public_key),
+            },
+        ))
+        .emit();
+
+        true
+    }
+
+    #[inline]
+    #[must_use]
+    fn maybe_add_public_key(&mut self, me: &AccountIdRef, public_key: PublicKey) -> bool {
+        if me == public_key.to_implicit_account_id() {
+            let was_removed = self.implicit_public_key_removed;
+            self.implicit_public_key_removed = false;
+            was_removed
+        } else {
+            self.public_keys.insert(public_key)
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn remove_public_key(&mut self, me: &AccountIdRef, public_key: &PublicKey) -> bool {
+        if !self.maybe_remove_public_key(me, public_key) {
+            return false;
+        }
+
+        DefuseEvent::PublicKeyRemoved(AccountEvent::new(
+            Cow::Borrowed(me),
+            PublicKeyEvent {
+                public_key: Cow::Borrowed(public_key),
+            },
+        ))
+        .emit();
+
+        true
+    }
+
+    #[inline]
+    #[must_use]
+    fn maybe_remove_public_key(&mut self, me: &AccountIdRef, public_key: &PublicKey) -> bool {
+        if me == public_key.to_implicit_account_id() {
+            let was_removed = self.implicit_public_key_removed;
+            self.implicit_public_key_removed = true;
+            !was_removed
+        } else {
+            self.public_keys.remove(public_key)
+        }
+    }
+
+    #[inline]
+    pub fn commit_nonce(&mut self, n: U256) -> Result<()> {
+        self.nonces.commit(n)
     }
 }
